requirements.txt
streamlit==1.40.0
plotly==5.24.1
pandas==2.2.3
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
Pillow==11.0.0
PyYAML==6.0.2
